requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
flask-crontab == 0.1.2
Flask-Migrate==3.0.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.1
PyJWT==2.1.0